Mere test

+++ blink.c
@@ -1,5 +1,5 @@
 
-int x = 0;
+uint8_t x = 0;
 while (x < 10){
 echo "Test lol";
 sleep(1);
